make wrappers LoopTimes (RuntimeFreq using run all times) to remove redundant run_at_time functions in each wrapper, ci-run-all-diff
